XYData and proof of concept of flow control
1. created class XYData to store data from each channel so if some data
doesn't come in on a point it has its own xdata and doesn't throw an
error when trying to plot. 2. added unit testing to pH class because it
gives data in either mS or microS. 3. proof of concept flow control
where parameters were preset in acquisition class.

Next steps: build GUI to enter parameters for acquisition class, figure
out best way to store and export the data that is acquired