added photon counter panel and variable number of plots
added panel to select channels, interval and dwell time for photon counter settings. also added other command line to functions to set gates, etc. changed plotting and writing data so it only plots/writes that data selected to be generated. has not been tested with pH meter and ADC so may be bugs.